requirements.txt
django
django-mptt
django-debug-toolbar
